projects-api/src/admin/auth.py
import time
from datetime import datetime

import jwt
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer

from config import config

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/admin/auth")


def create_access_token(user: str) -> str:
    payload = {"user": user, "expires": time.time() + 3600 * 10}
    token = jwt.encode(payload, config.JWT_SECRET_KEY, algorithm="HS256")
    return token


def verify_access_token(token: str) -> dict:
    try:
        data = jwt.decode(token, config.JWT_SECRET_KEY, algorithms=["HS256"])
        expire = data.get("expires")

        if expire is None:
            raise HTTPException(
                status_code=status.HTTP_400_BAD_REQUEST,
                detail="No access token suplied.",
            )
        if datetime.utcnow() > datetime.utcfromtimestamp(expire):
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN, detail="Token expired!"
            )

        return data

    except (
        jwt.exceptions.InvalidSignatureError,
        jwt.exceptions.DecodeError,
        jwt.exceptions.InvalidAlgorithmError,
    ):
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid token."
        )


async def authenticate(token: str = Depends(oauth2_scheme)) -> str:
    if not token:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN, detail="Sign in for access."
        )

    decoded_token = verify_access_token(token)
    if decoded_token["user"] != config.ADMIN_USERNAME:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Only admin can access to this event.",
        )

    return decoded_token["user"]
